Major cleanup for PUBDEV-291, HEXDEV-10 and HEXDEV-6.

+++ h2o-algos/src/test/java/hex/tree/drf/DRFTest.java
@@ -628,6 +628,7 @@ public class DRFTest extends TestUtil {
   @Test
   public void testNoRowWeights() {
     Frame tfr = null, vfr = null;
+    DRFModel drf = null;
 
     Scope.enter();
     try {
@@ -643,7 +644,7 @@ public class DRFTest extends TestUtil {
 
       // Build a first model; all remaining models should be equal
       DRF job = new DRF(parms);
-      DRFModel drf = job.trainModel().get();
+      drf = job.trainModel().get();
 
       drf.score(parms.train());
       hex.ModelMetricsBinomial mm = hex.ModelMetricsBinomial.getFromDKV(drf, parms.train());
@@ -653,18 +654,18 @@ public class DRFTest extends TestUtil {
       assertEquals(0.07692307692307693, mse, 1e-8);
 
       job.remove();
-      drf.delete();
     } finally {
       if (tfr != null) tfr.remove();
       if (vfr != null) vfr.remove();
+      if (drf != null) drf.remove();
+      Scope.exit();
     }
-    Scope.exit();
   }
 
-  @Ignore
   @Test
   public void testRowWeightsOne() {
     Frame tfr = null, vfr = null;
+    DRFModel drf = null;
 
     Scope.enter();
     try {
@@ -681,7 +682,7 @@ public class DRFTest extends TestUtil {
 
       // Build a first model; all remaining models should be equal
       DRF job = new DRF(parms);
-      DRFModel drf = job.trainModel().get();
+      drf = job.trainModel().get();
 
       drf.score(parms.train());
       hex.ModelMetricsBinomial mm = hex.ModelMetricsBinomial.getFromDKV(drf, parms.train());
@@ -690,17 +691,18 @@ public class DRFTest extends TestUtil {
       double mse = drf._output._training_metrics.mse();
       assertEquals(0.07692307692307693, mse, 1e-8); //Note: better results than non-shuffled
       job.remove();
-      drf.delete();
     } finally {
       if (tfr != null) tfr.remove();
       if (vfr != null) vfr.remove();
+      if (drf != null) drf.delete();
+      Scope.exit();
     }
-    Scope.exit();
   }
 
   @Test
   public void testNoRowWeightsShuffled() {
     Frame tfr = null, vfr = null;
+    DRFModel drf = null;
 
     Scope.enter();
     try {
@@ -716,7 +718,7 @@ public class DRFTest extends TestUtil {
 
       // Build a first model; all remaining models should be equal
       DRF job = new DRF(parms);
-      DRFModel drf = job.trainModel().get();
+      drf = job.trainModel().get();
 
       drf.score(parms.train());
       hex.ModelMetricsBinomial mm = hex.ModelMetricsBinomial.getFromDKV(drf, parms.train());
@@ -725,18 +727,19 @@ public class DRFTest extends TestUtil {
       double mse = drf._output._training_metrics.mse();
       assertEquals(0.11538629999502548, mse, 1e-8); //different rows are sampled -> results differ from unshuffled data
       job.remove();
-      drf.delete();
     } finally {
       if (tfr != null) tfr.remove();
       if (vfr != null) vfr.remove();
+      if (drf != null) drf.delete();
+      Scope.exit();
     }
-    Scope.exit();
   }
 
   @Ignore
   @Test
   public void testRowWeights() {
     Frame tfr = null, vfr = null;
+    DRFModel drf = null;
 
     Scope.enter();
     try {
@@ -753,7 +756,7 @@ public class DRFTest extends TestUtil {
 
       // Build a first model; all remaining models should be equal
       DRF job = new DRF(parms);
-      DRFModel drf = job.trainModel().get();
+      drf = job.trainModel().get();
 
       drf.score(parms.train());
       hex.ModelMetricsBinomial mm = hex.ModelMetricsBinomial.getFromDKV(drf, parms.train());
@@ -762,11 +765,11 @@ public class DRFTest extends TestUtil {
       double mse = drf._output._training_metrics.mse();
       assertEquals(0.07692307692307693, mse, 1e-8);
       job.remove();
-      drf.delete();
     } finally {
       if (tfr != null) tfr.remove();
       if (vfr != null) vfr.remove();
+      if (drf != null) drf.delete();
+      Scope.exit();
     }
-    Scope.exit();
   }
 }
